refactor: added separate files for modal styling and types
